PI-1285 Move component to a separate package
