+++ .github/workflows/build.yml
@@ -1,0 +1,26 @@
+name: Build component
+on:
+  push:
+    branches-ignore:
+      - main
+    paths:
+      - 'packages/probation-search/**'
+
+jobs:
+  build:
+    name: Build
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          cache: npm
+          node-version-file: .nvmrc
+
+      - name: Install dependencies
+        run: npm ci -w '@ministryofjustice/probation-search-frontend'
+
+      - name: Build
+        run: npm run build -w '@ministryofjustice/probation-search-frontend'
