+++ package.json
@@ -101,7 +101,6 @@
     "connect-flash": "^0.1.1",
     "connect-redis": "^7.1.0",
     "csurf": "^1.11.0",
-    "date-fns": "^2.30.0",
     "express": "^4.18.2",
     "express-prom-bundle": "^6.6.0",
     "express-session": "^1.17.3",
@@ -174,5 +173,8 @@
     "cypress": {
       "@cypress/request": "^3.0.0"
     }
-  }
+  },
+  "workspaces": [
+    "packages/probation-search"
+  ]
 }
