+++ packages/probation-search/.eslintignore
@@ -1,0 +1,6 @@
+node_modules
+public
+assets
+cypress.json
+reporter-config.json
+dist/
