+++ server/routes/index.ts
@@ -1,11 +1,10 @@
 import { type RequestHandler, Router } from 'express'
 
+import probationSearchRoutes from '@ministryofjustice/probation-search-frontend/dist/routes/search'
 import asyncMiddleware from '../middleware/asyncMiddleware'
 import config from '../config'
 import type { Services } from '../services'
-import probationSearchRoutes from '../component/routes/routes'
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
 export default function routes(service: Services): Router {
   const router = Router()
   const get = (path: string | string[], handler: RequestHandler) => router.get(path, asyncMiddleware(handler))
