+++ server/utils/nunjucksSetup.ts
@@ -32,6 +32,7 @@ export default function nunjucksSetup(app: express.Express, applicationInfo: App
       'node_modules/govuk-frontend/components/',
       'node_modules/@ministryofjustice/frontend/',
       'node_modules/@ministryofjustice/frontend/moj/components/',
+      'node_modules/@ministryofjustice/probation-search-frontend/dist/components',
     ],
     {
       autoescape: true,
